third_party/blink/web_tests/platform/generic/fast/events/pageshow-pagehide-expected.txt

Test pageshow/pagehide event behavior in subframes.

***Original load - onload and pageshow events should fire for subframes, and then for main frame***
Subsubframe window.onload
Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
Subframe window.onload
Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
Main frame window.onload
Main frame window.onpageshow, target = [object HTMLDocument], persisted = false
***Navigating bottom-level subframe, onpagehide events should fire for subsubframe***
Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subsubframe window.onunload
Subsubframe window.onload
Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
***Navigating mid-level subframe, onpagehide events should fire for both subframes***
Subframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subframe window.onunload
Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = false
Subsubframe window.onunload
Subframe window.onload
Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
***Done***
